app/requirements.txt
aiohttp==3.8.4
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.38.3
asyncpg==0.27.0
python-dotenv==1.0.0
pydantic==1.10.2
cachetools
bcrypt
email-validator
gunicorn==20.1.0
pylint==2.17.1
psycopg2-binary==2.9.5